Revert "Spark 3.5 Support (#39395)"
This reverts commit e9854e5b2eb22d734544669b36031fbbcc116f6a.
