+++ .vscode/cspell.json
@@ -136,7 +136,6 @@
     "sdk/cosmos/azure-cosmos-spark_3-2_2-12/**",
     "sdk/cosmos/azure-cosmos-spark_3-3_2-12/**",
     "sdk/cosmos/azure-cosmos-spark_3-4_2-12/**",
-    "sdk/cosmos/azure-cosmos-spark_3-5_2-12/**",
     "sdk/core/azure-core-test/**",
     "sdk/cosmos/azure-cosmos-encryption/**",
     "sdk/cosmos/azure-cosmos-spark_3_2-12/**",
