+++ eng/versioning/version_client.txt
@@ -109,7 +109,6 @@ com.azure.cosmos.spark:azure-cosmos-spark_3-1_2-12;4.28.4;4.29.0-beta.1
 com.azure.cosmos.spark:azure-cosmos-spark_3-2_2-12;4.28.4;4.29.0-beta.1
 com.azure.cosmos.spark:azure-cosmos-spark_3-3_2-12;4.28.4;4.29.0-beta.1
 com.azure.cosmos.spark:azure-cosmos-spark_3-4_2-12;4.28.4;4.29.0-beta.1
-com.azure.cosmos.spark:azure-cosmos-spark_3-5_2-12;4.29.0-beta.1;4.29.0-beta.1
 com.azure:azure-cosmos-encryption;2.9.0;2.10.0-beta.1
 com.azure:azure-cosmos-test;1.0.0-beta.6;1.0.0-beta.7
 com.azure:azure-cosmos-tests;1.0.0-beta.1;1.0.0-beta.1
