--- sdk/cosmos/azure-cosmos-spark_3-5_2-12/src/main/resources/azure-cosmos-spark.properties
@@ -1,2 +1,0 @@
-name=${project.artifactId}
-version=${project.version}
