--- sdk/cosmos/azure-cosmos-spark_3-5_2-12/src/main/scala/com/azure/cosmos/spark/CosmosRecordsWrittenMetric.scala
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-package com.azure.cosmos.spark
-
-import org.apache.spark.sql.connector.metric.CustomSumMetric
-
-private[cosmos] class CosmosRecordsWrittenMetric extends CustomSumMetric {
-  override def name(): String = CosmosConstants.MetricNames.RecordsWritten
-
-  override def description(): String = CosmosConstants.MetricNames.RecordsWritten
-}
