+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/main/scala/com/azure/cosmos/spark/ItemsBatchWriter.scala
@@ -40,11 +40,6 @@ private class ItemsBatchWriter
       sparkEnvironmentInfo)
   }
 
-  override def useCommitCoordinator(): Boolean = {
-    return false
-    // TODO
-  }
-
   override def commit(writerCommitMessages: Array[WriterCommitMessage]): Unit = {
     // TODO
   }
