+++ sdk/cosmos/azure-cosmos-spark_3_2-12/src/test/scala/com/azure/cosmos/spark/RowSerializerPollSpec.scala
@@ -3,9 +3,10 @@
 package com.azure.cosmos.spark
 
 import com.azure.cosmos.spark.diagnostics.BasicLoggingTrait
+import org.apache.spark.sql.catalyst.encoders.RowEncoder
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
 
-abstract class RowSerializerPollSpec extends UnitSpec with BasicLoggingTrait {
+class RowSerializerPollSpec extends UnitSpec with BasicLoggingTrait {
   //scalastyle:off multiple.string.literals
 
   "RowSerializer returned to the pool" should "be reused when capacity not exceeded" in {
@@ -23,6 +24,21 @@ abstract class RowSerializerPollSpec extends UnitSpec with BasicLoggingTrait {
 
     val newSerializer = RowSerializerPool.getOrCreateSerializer(sameSchema)
     serializer.eq(newSerializer) shouldBe false
+  }
+
+  "RowSerializer " should "be returned to the pool only a limited number of times" in {
+    val canRun = Platform.canRunTestAccessingDirectByteBuffer
+    assume(canRun._1, canRun._2)
+
+    val schema = StructType(Seq(StructField("column01", IntegerType), StructField("column02", StringType)))
+
+    for (_ <- 1 to 256) {
+      RowSerializerPool.returnSerializerToPool(schema, RowEncoder(schema).createSerializer()) shouldBe true
+    }
+
+    logInfo("First 256 attempt to pool succeeded")
+
+    RowSerializerPool.returnSerializerToPool(schema, RowEncoder(schema).createSerializer()) shouldBe false
   }
   //scalastyle:on multiple.string.literals
 }
