+++ sdk/cosmos/ci.yml
@@ -19,7 +19,6 @@ trigger:
       - sdk/cosmos/azure-cosmos-spark_3-2_2-12/
       - sdk/cosmos/azure-cosmos-spark_3-3_2-12/
       - sdk/cosmos/azure-cosmos-spark_3-4_2-12/
-      - sdk/cosmos/azure-cosmos-spark_3-5_2-12/
       - sdk/cosmos/azure-cosmos-test/
       - sdk/cosmos/azure-cosmos-tests/
       - sdk/cosmos/azure-cosmos-kafka-connect/
@@ -36,7 +35,6 @@ trigger:
       - sdk/cosmos/azure-cosmos-spark_3-2_2-12/pom.xml
       - sdk/cosmos/azure-cosmos-spark_3-3_2-12/pom.xml
       - sdk/cosmos/azure-cosmos-spark_3-4_2-12/pom.xml
-      - sdk/cosmos/azure-cosmos-spark_3-5_2-12/pom.xml
       - sdk/cosmos/azure-cosmos-kafka-connect/pom.xml
 
 pr:
@@ -62,7 +60,6 @@ pr:
       - sdk/cosmos/azure-cosmos-spark_3-2_2-12/
       - sdk/cosmos/azure-cosmos-spark_3-3_2-12/
       - sdk/cosmos/azure-cosmos-spark_3-4_2-12/
-      - sdk/cosmos/azure-cosmos-spark_3-5_2-12/
       - sdk/cosmos/faq/
       - sdk/cosmos/azure-cosmos-kafka-connect/
     exclude:
@@ -76,7 +73,6 @@ pr:
       - sdk/cosmos/azure-cosmos-spark_3-2_2-12/pom.xml
       - sdk/cosmos/azure-cosmos-spark_3-3_2-12/pom.xml
       - sdk/cosmos/azure-cosmos-spark_3-4_2-12/pom.xml
-      - sdk/cosmos/azure-cosmos-spark_3-5_2-12/pom.xml
       - sdk/cosmos/azure-cosmos-test/pom.xml
       - sdk/cosmos/azure-cosmos-tests/pom.xml
       - sdk/cosmos/azure-cosmos-kafka-connect/pom.xml
@@ -103,10 +99,6 @@ parameters:
     default: true
   - name: release_azurecosmosspark34
     displayName: 'azure-cosmos-spark_3-4_2-12'
-    type: boolean
-    default: true
-  - name: release_azurecosmosspark35
-    displayName: 'azure-cosmos-spark_3-5_2-12'
     type: boolean
     default: true
   - name: release_azurecosmostest
@@ -161,13 +153,6 @@ extends:
         skipPublishDocGithubIo: true
         skipPublishDocMs: true
         releaseInBatch: ${{ parameters.release_azurecosmosspark34 }}
-      - name: azure-cosmos-spark_3-5_2-12
-        groupId: com.azure.cosmos.spark
-        safeName: azurecosmosspark35
-        uberJar: true
-        skipPublishDocGithubIo: true
-        skipPublishDocMs: true
-        releaseInBatch: ${{ parameters.release_azurecosmosspark35 }}
       - name: azure-cosmos-test
         groupId: com.azure
         safeName: azurecosmostest
