+++ sdk/cosmos/pom.xml
@@ -19,7 +19,6 @@
     <module>azure-cosmos-spark_3-2_2-12</module>
     <module>azure-cosmos-spark_3-3_2-12</module>
     <module>azure-cosmos-spark_3-4_2-12</module>
-    <module>azure-cosmos-spark_3-5_2-12</module>
     <module>azure-cosmos-test</module>
     <module>azure-cosmos-tests</module>
     <module>azure-cosmos-kafka-connect</module>
